Fix Hosted operation for Windows (#2378)
This PR updates the Hosted Component so it builds and runs for Windows and also Esp32 architecture.

I've done some basic testing with Hosted running on an ESP32-S2 target, which pretty much worked straight away (apart from `pulseIn` signature). However, as I use WSL if using a serial interface the Host application must be run in a Windows build. That needed a few fixes.

This also seems like a good time to get the Kconfig settings for this Component updated!

* Move `simpleRPC` into Components (Hosted is a Component so should not depend on a Library)
* Fix compilation problems for Windows
* Update environment variable handling & Kconfig settings
* Fix `pulseIn` signature for Esp32

+++ Sming/Arch/Esp32/Core/Digital.cpp
@@ -127,7 +127,7 @@ void noPullup(uint16_t pin)
 		}                                                                                                              \
 	}
 
-unsigned long pulseIn(uint8_t pin, uint8_t state, unsigned long timeout)
+unsigned long pulseIn(uint16_t pin, uint8_t state, unsigned long timeout)
 {
 	const uint32_t max_timeout_us = clockCyclesToMicroseconds(UINT_MAX);
 	if(timeout > max_timeout_us) {
